+++ pydl/pydlutils/tests/test_sdss.py
@@ -131,6 +131,8 @@ class TestSDSS(object):
         assert (np.array([1237661382772195474, 1237649770790322299],
                          dtype=np.int64) ==
                 sdss_objid(run, camcol, field, obj)).all()
+        assert (sdss_objid(752, 5, 618, 459, rerun=40,
+                           skyversion=1, firstfield=1) == 587722984180548043)
         #
         # Exceptions
         #
@@ -147,9 +149,15 @@ class TestSDSS(object):
         with raises(ValueError):
             objid = sdss_objid(3704, 3, 91, 146, skyversion=np.array([2, 3]))
         with raises(ValueError):
+            objid = sdss_objid(3704, 3, 91, 146, firstfield=np.array([0, 1]))
+        with raises(ValueError):
             objid = sdss_objid(3704, 3, 91, 146, skyversion=-2)
         with raises(ValueError):
             objid = sdss_objid(3704, 3, 91, 146, skyversion=16)
+        with raises(ValueError):
+            objid = sdss_objid(3704, 3, 91, 146, firstfield=-2)
+        with raises(ValueError):
+            objid = sdss_objid(3704, 3, 91, 146, firstfield=2)
         with raises(ValueError):
             objid = sdss_objid(3704, 3, 91, 146, rerun=-2)
         with raises(ValueError):
@@ -190,6 +198,21 @@ class TestSDSS(object):
         #
         # Exceptions
         #
+        with raises(ValueError):
+            s = sdss_specobjid(np.array([4055, 4056]), 408, 55359, 'v5_7_0')
+        with raises(ValueError):
+            s = sdss_specobjid(4055, np.array([408, 409]), 55359, 'v5_7_0')
+        with raises(ValueError):
+            s = sdss_specobjid(4055, 408, np.array([55359, 55360]), 'v5_7_0')
+        with raises(ValueError):
+            s = sdss_specobjid(4055, 408, 55359,
+                               np.array(['v5_7_0', 'v5_7_2']))
+        with raises(ValueError):
+            s = sdss_specobjid(4055, 408, 55359, 'v5_7_0',
+                               line=np.array([137, 138]))
+        with raises(ValueError):
+            s = sdss_specobjid(4055, 408, 55359, 'v5_7_0',
+                               index=np.array([137, 138]))
         with raises(ValueError):
             s = sdss_specobjid(4055, 408, 55359, 'v5_7_0', line=137,
                                index=137)